Add RAG pipeline with pgvector embeddings for intelligent lesson planning
