+++ app/lib/lessonPlan/constants.ts
@@ -1,0 +1,657 @@
+import { Curriculum, Source } from "@/app/types/lessonPlan";
+
+export type GradeCategory = "early" | "elementary" | "middle" | "high";
+
+// Grade level mappings
+export const gradeLevelMap: { [key: string]: string } = {
+  "1": "GRADE_1",
+  "2": "GRADE_2",
+  "3": "GRADE_3",
+  "4": "GRADE_4",
+  "5": "GRADE_5",
+  "6": "GRADE_6",
+  "7": "GRADE_7",
+  "8": "GRADE_8",
+  "9": "GRADE_9",
+  "10": "GRADE_10",
+  "11": "GRADE_11",
+  "12": "GRADE_12",
+  infant: "INFANT",
+  toddler: "TODDLER",
+  preschool: "PRESCHOOL",
+  kindergarten: "KINDERGARTEN",
+  grade_1: "GRADE_1",
+  grade_2: "GRADE_2",
+  grade_3: "GRADE_3",
+  grade_4: "GRADE_4",
+  grade_5: "GRADE_5",
+  grade_6: "GRADE_6",
+  grade_7: "GRADE_7",
+  grade_8: "GRADE_8",
+  grade_9: "GRADE_9",
+  grade_10: "GRADE_10",
+  grade_11: "GRADE_11",
+  grade_12: "GRADE_12",
+  "Grade 1": "GRADE_1",
+  "Grade 2": "GRADE_2",
+  "Grade 3": "GRADE_3",
+  "Grade 4": "GRADE_4",
+  "Grade 5": "GRADE_5",
+  "Grade 6": "GRADE_6",
+  "Grade 7": "GRADE_7",
+  "Grade 8": "GRADE_8",
+  "Grade 9": "GRADE_9",
+  "Grade 10": "GRADE_10",
+  "Grade 11": "GRADE_11",
+  "Grade 12": "GRADE_12",
+  INFANT: "INFANT",
+  TODDLER: "TODDLER",
+  PRESCHOOL: "PRESCHOOL",
+  KINDERGARTEN: "KINDERGARTEN",
+  GRADE_1: "GRADE_1",
+  GRADE_2: "GRADE_2",
+  GRADE_3: "GRADE_3",
+  GRADE_4: "GRADE_4",
+  GRADE_5: "GRADE_5",
+  GRADE_6: "GRADE_6",
+  GRADE_7: "GRADE_7",
+  GRADE_8: "GRADE_8",
+  GRADE_9: "GRADE_9",
+  GRADE_10: "GRADE_10",
+  GRADE_11: "GRADE_11",
+  GRADE_12: "GRADE_12",
+};
+
+export const validGradeLevels = [
+  "INFANT",
+  "TODDLER",
+  "PRESCHOOL",
+  "KINDERGARTEN",
+  "GRADE_1",
+  "GRADE_2",
+  "GRADE_3",
+  "GRADE_4",
+  "GRADE_5",
+  "GRADE_6",
+  "GRADE_7",
+  "GRADE_8",
+  "GRADE_9",
+  "GRADE_10",
+  "GRADE_11",
+  "GRADE_12",
+];
+
+export const earlyGrades = ["INFANT", "TODDLER", "PRESCHOOL", "KINDERGARTEN"];
+export const elementaryGrades = [
+  "GRADE_1",
+  "GRADE_2",
+  "GRADE_3",
+  "GRADE_4",
+  "GRADE_5",
+];
+export const middleSchoolGrades = ["GRADE_6", "GRADE_7", "GRADE_8"];
+export const highSchoolGrades = ["GRADE_9", "GRADE_10", "GRADE_11", "GRADE_12"];
+
+// Subject mappings
+export const subjectMap: { [key: string]: string } = {
+  literacy: "LITERACY",
+  math: "MATH",
+  science: "SCIENCE",
+  Science: "SCIENCE",
+  art: "ART",
+  music: "MUSIC",
+  physical_education: "PHYSICAL_EDUCATION",
+  social_emotional: "SOCIAL_EMOTIONAL",
+  history: "HISTORY",
+  geography: "GEOGRAPHY",
+  stem: "STEM",
+  foreign_language: "FOREIGN_LANGUAGE",
+  computer_science: "COMPUTER_SCIENCE",
+  "computer science": "COMPUTER_SCIENCE",
+  "Computer Science": "COMPUTER_SCIENCE",
+  civics: "CIVICS",
+  english: "ENGLISH",
+  mathematics: "MATHEMATICS",
+  the_arts: "THE_ARTS",
+  health_pe: "HEALTH_PE",
+  humanities_social_sciences: "HUMANITIES_SOCIAL_SCIENCES",
+  technologies: "TECHNOLOGIES",
+  languages: "LANGUAGES",
+  civics_citizenship: "CIVICS_CITIZENSHIP",
+  sensory_development: "SENSORY_DEVELOPMENT",
+  fine_motor_skills: "FINE_MOTOR_SKILLS",
+  language_development: "LANGUAGE_DEVELOPMENT",
+  social_studies: "SOCIAL_STUDIES",
+  drama: "DRAMA",
+  dance: "DANCE",
+  health_and_wellness: "HEALTH_AND_WELLNESS",
+  character_education: "CHARACTER_EDUCATION",
+  community_service: "COMMUNITY_SERVICE",
+  engineering: "ENGINEERING",
+  business: "BUSINESS",
+  economics: "ECONOMICS",
+  philosophy: "PHILOSOPHY",
+  LITERACY: "LITERACY",
+  MATH: "MATH",
+  SCIENCE: "SCIENCE",
+  ART: "ART",
+  MUSIC: "MUSIC",
+  PHYSICAL_EDUCATION: "PHYSICAL_EDUCATION",
+  SOCIAL_EMOTIONAL: "SOCIAL_EMOTIONAL",
+  HISTORY: "HISTORY",
+  GEOGRAPHY: "GEOGRAPHY",
+  STEM: "STEM",
+  FOREIGN_LANGUAGE: "FOREIGN_LANGUAGE",
+  COMPUTER_SCIENCE: "COMPUTER_SCIENCE",
+  CIVICS: "CIVICS",
+  ENGLISH: "ENGLISH",
+  MATHEMATICS: "MATHEMATICS",
+  THE_ARTS: "THE_ARTS",
+  HEALTH_PE: "HEALTH_PE",
+  HUMANITIES_SOCIAL_SCIENCES: "HUMANITIES_SOCIAL_SCIENCES",
+  TECHNOLOGIES: "TECHNOLOGIES",
+  LANGUAGES: "LANGUAGES",
+  CIVICS_CITIZENSHIP: "CIVICS_CITIZENSHIP",
+  SENSORY_DEVELOPMENT: "SENSORY_DEVELOPMENT",
+  FINE_MOTOR_SKILLS: "FINE_MOTOR_SKILLS",
+  LANGUAGE_DEVELOPMENT: "LANGUAGE_DEVELOPMENT",
+  SOCIAL_STUDIES: "SOCIAL_STUDIES",
+  DRAMA: "DRAMA",
+  DANCE: "DANCE",
+  HEALTH_AND_WELLNESS: "HEALTH_AND_WELLNESS",
+  CHARACTER_EDUCATION: "CHARACTER_EDUCATION",
+  COMMUNITY_SERVICE: "COMMUNITY_SERVICE",
+  ENGINEERING: "ENGINEERING",
+  BUSINESS: "BUSINESS",
+  ECONOMICS: "ECONOMICS",
+  PHILOSOPHY: "PHILOSOPHY",
+};
+
+export const validSubjectsList = [
+  "LITERACY",
+  "MATH",
+  "SCIENCE",
+  "ART",
+  "MUSIC",
+  "PHYSICAL_EDUCATION",
+  "SOCIAL_EMOTIONAL",
+  "HISTORY",
+  "GEOGRAPHY",
+  "STEM",
+  "FOREIGN_LANGUAGE",
+  "COMPUTER_SCIENCE",
+  "CIVICS",
+  "ENGLISH",
+  "MATHEMATICS",
+  "THE_ARTS",
+  "HEALTH_PE",
+  "HUMANITIES_SOCIAL_SCIENCES",
+  "TECHNOLOGIES",
+  "LANGUAGES",
+  "CIVICS_CITIZENSHIP",
+  "SENSORY_DEVELOPMENT",
+  "FINE_MOTOR_SKILLS",
+  "LANGUAGE_DEVELOPMENT",
+  "SOCIAL_STUDIES",
+  "DRAMA",
+  "DANCE",
+  "HEALTH_AND_WELLNESS",
+  "CHARACTER_EDUCATION",
+  "COMMUNITY_SERVICE",
+  "ENGINEERING",
+  "BUSINESS",
+  "ECONOMICS",
+  "PHILOSOPHY",
+];
+
+// Theme mappings
+export const themeMap: { [key: string]: string } = {
+  seasons: "SEASONS",
+  nature: "NATURE",
+  holidays: "HOLIDAYS",
+  emotions: "EMOTIONS",
+  community: "COMMUNITY",
+  animals: "ANIMALS",
+  transportation: "TRANSPORTATION",
+  colors: "COLORS",
+  colours: "COLOURS",
+  shapes: "SHAPES",
+  numbers: "NUMBERS",
+  culture: "CULTURE",
+  history: "HISTORY",
+  science_fiction: "SCIENCE_FICTION",
+  technology: "TECHNOLOGY",
+  global_issues: "GLOBAL_ISSUES",
+  literature: "LITERATURE",
+  indigenous_culture: "INDIGENOUS_CULTURE",
+  australian_history: "AUSTRALIAN_HISTORY",
+  sustainability: "SUSTAINABILITY",
+  transport: "TRANSPORT",
+  SEASONS: "SEASONS",
+  NATURE: "NATURE",
+  HOLIDAYS: "HOLIDAYS",
+  EMOTIONS: "EMOTIONS",
+  COMMUNITY: "COMMUNITY",
+  ANIMALS: "ANIMALS",
+  TRANSPORTATION: "TRANSPORTATION",
+  COLORS: "COLORS",
+  COLOURS: "COLOURS",
+  SHAPES: "SHAPES",
+  NUMBERS: "NUMBERS",
+  CULTURE: "CULTURE",
+  HISTORY: "HISTORY",
+  SCIENCE_FICTION: "SCIENCE_FICTION",
+  TECHNOLOGY: "TECHNOLOGY",
+  GLOBAL_ISSUES: "GLOBAL_ISSUES",
+  LITERATURE: "LITERATURE",
+  INDIGENOUS_CULTURE: "INDIGENOUS_CULTURE",
+  AUSTRALIAN_HISTORY: "AUSTRALIAN_HISTORY",
+  SUSTAINABILITY: "SUSTAINABILITY",
+  TRANSPORT: "TRANSPORT",
+};
+
+export const validThemesList = [
+  "SEASONS",
+  "NATURE",
+  "HOLIDAYS",
+  "EMOTIONS",
+  "COMMUNITY",
+  "ANIMALS",
+  "TRANSPORTATION",
+  "COLORS",
+  "COLOURS",
+  "SHAPES",
+  "NUMBERS",
+  "CULTURE",
+  "HISTORY",
+  "SCIENCE_FICTION",
+  "TECHNOLOGY",
+  "GLOBAL_ISSUES",
+  "LITERATURE",
+  "INDIGENOUS_CULTURE",
+  "AUSTRALIAN_HISTORY",
+  "SUSTAINABILITY",
+  "TRANSPORT",
+];
+
+// Valid subjects by curriculum and grade category
+export const validSubjects: Record<Curriculum, Record<GradeCategory, string[]>> =
+  {
+    US: {
+      early: [
+        "LITERACY",
+        "MATH",
+        "SCIENCE",
+        "ART",
+        "MUSIC",
+        "PHYSICAL_EDUCATION",
+        "SOCIAL_EMOTIONAL",
+      ],
+      elementary: [
+        "LITERACY",
+        "MATH",
+        "SCIENCE",
+        "ART",
+        "MUSIC",
+        "PHYSICAL_EDUCATION",
+        "SOCIAL_EMOTIONAL",
+        "HISTORY",
+        "GEOGRAPHY",
+      ],
+      middle: [
+        "LITERACY",
+        "MATH",
+        "SCIENCE",
+        "ART",
+        "MUSIC",
+        "PHYSICAL_EDUCATION",
+        "SOCIAL_EMOTIONAL",
+        "HISTORY",
+        "LITERATURE",
+        "GEOGRAPHY",
+        "STEM",
+      ],
+      high: [
+        "LITERACY",
+        "MATH",
+        "SCIENCE",
+        "ART",
+        "MUSIC",
+        "PHYSICAL_EDUCATION",
+        "SOCIAL_EMOTIONAL",
+        "HISTORY",
+        "LITERATURE",
+        "GEOGRAPHY",
+        "STEM",
+        "FOREIGN_LANGUAGE",
+        "COMPUTER_SCIENCE",
+        "CIVICS",
+      ],
+    },
+    AUS: {
+      early: [
+        "ENGLISH",
+        "MATHEMATICS",
+        "SCIENCE",
+        "THE_ARTS",
+        "HEALTH_PE",
+        "SOCIAL_EMOTIONAL",
+      ],
+      elementary: [
+        "ENGLISH",
+        "MATHEMATICS",
+        "SCIENCE",
+        "THE_ARTS",
+        "HEALTH_PE",
+        "SOCIAL_EMOTIONAL",
+        "HUMANITIES_SOCIAL_SCIENCES",
+        "TECHNOLOGIES",
+      ],
+      middle: [
+        "ENGLISH",
+        "MATHEMATICS",
+        "SCIENCE",
+        "THE_ARTS",
+        "HEALTH_PE",
+        "SOCIAL_EMOTIONAL",
+        "HUMANITIES_SOCIAL_SCIENCES",
+        "TECHNOLOGIES",
+        "LANGUAGES",
+      ],
+      high: [
+        "ENGLISH",
+        "MATHEMATICS",
+        "SCIENCE",
+        "THE_ARTS",
+        "HEALTH_PE",
+        "SOCIAL_EMOTIONAL",
+        "HUMANITIES_SOCIAL_SCIENCES",
+        "TECHNOLOGIES",
+        "LANGUAGES",
+        "CIVICS_CITIZENSHIP",
+      ],
+    },
+  };
+
+// Valid themes by curriculum and grade category
+export const validThemes: Record<Curriculum, Record<GradeCategory, string[]>> =
+  {
+    US: {
+      early: [
+        "SEASONS",
+        "NATURE",
+        "HOLIDAYS",
+        "EMOTIONS",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORTATION",
+        "COLORS",
+        "SHAPES",
+        "NUMBERS",
+      ],
+      elementary: [
+        "SEASONS",
+        "NATURE",
+        "HOLIDAYS",
+        "EMOTIONS",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORTATION",
+        "COLORS",
+        "SHAPES",
+        "NUMBERS",
+        "CULTURE",
+        "HISTORY",
+      ],
+      middle: [
+        "SEASONS",
+        "NATURE",
+        "HOLIDAYS",
+        "EMOTIONS",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORTATION",
+        "CULTURE",
+        "HISTORY",
+        "SCIENCE_FICTION",
+        "TECHNOLOGY",
+      ],
+      high: [
+        "SEASONS",
+        "NATURE",
+        "HOLIDAYS",
+        "EMOTIONS",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORTATION",
+        "CULTURE",
+        "HISTORY",
+        "SCIENCE_FICTION",
+        "GLOBAL_ISSUES",
+        "TECHNOLOGY",
+        "LITERATURE",
+      ],
+    },
+    AUS: {
+      early: [
+        "SEASONS",
+        "NATURE",
+        "INDIGENOUS_CULTURE",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORT",
+        "COLOURS",
+        "SHAPES",
+        "NUMBERS",
+      ],
+      elementary: [
+        "SEASONS",
+        "NATURE",
+        "INDIGENOUS_CULTURE",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORT",
+        "COLOURS",
+        "SHAPES",
+        "NUMBERS",
+        "AUSTRALIAN_HISTORY",
+        "SUSTAINABILITY",
+      ],
+      middle: [
+        "SEASONS",
+        "NATURE",
+        "INDIGENOUS_CULTURE",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORT",
+        "AUSTRALIAN_HISTORY",
+        "SUSTAINABILITY",
+        "TECHNOLOGY",
+      ],
+      high: [
+        "SEASONS",
+        "NATURE",
+        "INDIGENOUS_CULTURE",
+        "COMMUNITY",
+        "ANIMALS",
+        "TRANSPORT",
+        "AUSTRALIAN_HISTORY",
+        "SUSTAINABILITY",
+        "GLOBAL_ISSUES",
+        "TECHNOLOGY",
+        "LITERATURE",
+      ],
+    },
+  };
+
+// Valid activity types by curriculum and grade category
+export const validActivityTypes: Record<
+  Curriculum,
+  Record<GradeCategory, string[]>
+> = {
+  US: {
+    early: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "FREE_PLAY",
+      "OUTDOOR",
+    ],
+    elementary: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "EXPERIMENT",
+      "FREE_PLAY",
+      "OUTDOOR",
+      "WRITING",
+      "PROJECT",
+    ],
+    middle: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "EXPERIMENT",
+      "OUTDOOR",
+      "GROUP_DISCUSSION",
+      "PROJECT",
+      "PRESENTATION",
+      "WRITING",
+    ],
+    high: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "EXPERIMENT",
+      "OUTDOOR",
+      "GROUP_DISCUSSION",
+      "PROJECT",
+      "PRESENTATION",
+      "WRITING",
+      "RESEARCH",
+      "DEBATE",
+      "CODING",
+    ],
+  },
+  AUS: {
+    early: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "FREE_PLAY",
+      "OUTDOOR",
+      "INDIGENOUS_STORY",
+    ],
+    elementary: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "EXPERIMENT",
+      "FREE_PLAY",
+      "OUTDOOR",
+      "WRITING",
+      "PROJECT",
+      "INDIGENOUS_STORY",
+    ],
+    middle: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "EXPERIMENT",
+      "OUTDOOR",
+      "GROUP_DISCUSSION",
+      "PROJECT",
+      "PRESENTATION",
+      "WRITING",
+      "INDIGENOUS_STORY",
+    ],
+    high: [
+      "STORYTELLING",
+      "CRAFT",
+      "MOVEMENT",
+      "MUSIC",
+      "EXPERIMENT",
+      "OUTDOOR",
+      "GROUP_DISCUSSION",
+      "PROJECT",
+      "PRESENTATION",
+      "WRITING",
+      "RESEARCH",
+      "DEBATE",
+      "CODING",
+      "INDIGENOUS_STORY",
+    ],
+  },
+};
+
+// Valid curriculums
+export const validCurriculums: Curriculum[] = ["US", "AUS"];
+
+// Default sources by curriculum
+export const getDefaultSources = (curriculum: Curriculum): Source[] => {
+  return curriculum === "US"
+    ? [
+        {
+          name: "Khan Academy",
+          url: "https://www.khanacademy.org",
+          description: "Used for activity ideas and educational content",
+        },
+        {
+          name: "NSTA",
+          url: "https://www.nsta.org",
+          description: "Used for science education resources",
+        },
+      ]
+    : [
+        {
+          name: "ACARA",
+          url: "https://www.australiancurriculum.edu.au",
+          description:
+            "Used for Australian curriculum standards and resources",
+        },
+        {
+          name: "Scootle",
+          url: "https://www.scootle.edu.au",
+          description: "Used for Australian educational resources",
+        },
+      ];
+};
+
+// Valid standards frameworks by curriculum
+export const getValidStandardsFrameworks = (curriculum: Curriculum): string[] => {
+  return curriculum === "US"
+    ? ["COMMON_CORE", "NGSS", "STATE_SPECIFIC"]
+    : ["ACARA"];
+};
+
+// Source scores for citation
+export const sourceScores: { [key: string]: number } = {
+  "Common Core": 90,
+  NGSS: 90,
+  ACARA: 90,
+  "Khan Academy": 85,
+  NSTA: 80,
+  Scootle: 85,
+};
+
+// Helper function to get grade category
+export const getGradeCategory = (gradeLevel: string): GradeCategory => {
+  if (earlyGrades.includes(gradeLevel)) {
+    return "early";
+  } else if (elementaryGrades.includes(gradeLevel)) {
+    return "elementary";
+  } else if (middleSchoolGrades.includes(gradeLevel)) {
+    return "middle";
+  } else if (highSchoolGrades.includes(gradeLevel)) {
+    return "high";
+  } else {
+    throw new Error(`Invalid gradeLevel: ${gradeLevel}`);
+  }
+};
